server/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0
django-cors-headers==3.2.10